AKACommons: preliminary support for observable collections and dynamic data sources (not working yet), recreated project to get rid of umbrella header problems, minor cleanup

+++ AKACommons/AKACommons/Classes/AKAIPNetmask.m
@@ -22,7 +22,7 @@
     NSUInteger v = 0;
     if (length > 0)
     {
-        for (int i=0; i < length; ++i)
+        for (NSUInteger i=0; i < length; ++i)
         {
             v = (v << 1) | 1;
         }
